table/requirements.txt
anyio==4.3.0
asgiref==3.7.2
cached-property==1.5.2
Cerberus==1.3.5
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
confusable-homoglyphs==3.3.1
Django==4.2.9
django-multiupload==0.6.1
django-registration==3.4
djangorestframework==3.14.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
iniconfig==2.0.0
isort==5.13.2
kolo==2.21.1
msgpack==1.0.8
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
psycopg2==2.9.9
pytest==8.1.1
pytest-django==4.8.0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
sniffio==1.3.1
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
toolz==0.12.1
typing_extensions==4.9.0
tzdata==2023.4
ulid-py==1.1.0
Unipath==1.1
zipp==3.18.1
